requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django
django-autoslug==1.9.8
django-ckeditor==6.5.1
django-js-asset==2.0.0
idna==3.4
mysqlclient
Pillow==9.3.0
python-decouple
razorpay==1.3.0
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.13
whitenoise==6.2.0
